refactor(): granulated includes - reduce compile time
