+++ echo/src/module.cpp
@@ -1,5 +1,8 @@
 #include "cocaine/service/echo.hpp"
 
+#include <cocaine/repository.hpp>
+#include <cocaine/repository/service.hpp>
+
 extern "C" {
     auto validation() -> cocaine::api::preconditions_t {
         return cocaine::api::preconditions_t{COCAINE_MAKE_VERSION(0, 12, 0)};
