+++ elasticsearch/src/module.cpp
@@ -20,6 +20,9 @@
 
 #include "cocaine/service/elasticsearch.hpp"
 
+#include <cocaine/repository.hpp>
+#include <cocaine/repository/service.hpp>
+
 using namespace cocaine;
 using namespace cocaine::service;
