+++ graphite/src/module.cpp
@@ -15,6 +15,9 @@
 
 #include <cocaine/graphite.hpp>
 
+#include <cocaine/repository.hpp>
+#include <cocaine/repository/service.hpp>
+
 using namespace cocaine;
 using namespace cocaine::service;
