+++ node/src/module.cpp
@@ -1,4 +1,8 @@
 #include <cocaine/detail/isolate/external.hpp>
+#include <cocaine/repository.hpp>
+#include <cocaine/repository/service.hpp>
+
+#include "cocaine/repository/isolate.hpp"
 #include "cocaine/service/node.hpp"
 
 #include "cocaine/detail/isolate/process.hpp"
