+++ unicorn/src/unicorn/zookeeper/increment.cpp
@@ -15,12 +15,13 @@
 
 #include "cocaine/detail/unicorn/zookeeper/increment.hpp"
 
+#include "cocaine/detail/future.hpp"
 #include "cocaine/detail/zookeeper/errors.hpp"
 
 #include <blackhole/logger.hpp>
 
+#include <cocaine/errors.hpp>
 #include <cocaine/logging.hpp>
-#include <cocaine/detail/future.hpp>
 
 namespace cocaine { namespace unicorn {
