+++ unicorn/src/unicorn/zookeeper/subscribe.cpp
@@ -15,12 +15,14 @@
 
 #include "cocaine/detail/unicorn/zookeeper/subscribe.hpp"
 
+#include "cocaine/detail/future.hpp"
 #include "cocaine/detail/zookeeper/errors.hpp"
+
+#include <cocaine/errors.hpp>
+#include <cocaine/logging.hpp>
 
 #include <blackhole/logger.hpp>
 
-#include <cocaine/logging.hpp>
-#include <cocaine/detail/future.hpp>
 
 namespace cocaine { namespace unicorn {
